requirements.txt
Django==1.9.7
django-auditlog==0.3.3
django-filter==0.13.0
django-jsonfield==1.0.0
djangorestframework==3.4.0
Markdown==2.6.6
Pillow==3.2.0
rest-framework-generic-relations==1.1.0
wheel==0.24.0
